added a popup javascript modal to confirm if i want to delete category before deleting it

+++ admin/function.php
@@ -39,7 +39,8 @@ function readAllCategories(){
         echo "<td>{$admin_cat_id}</td>";
         echo "<td>{$admin_cat_title}</td>";
         echo "<td><a href='categories.php?edit={$admin_cat_id}'>Edit</a></td>";
-        echo "<td><a href='categories.php?delete={$admin_cat_id}'>Delete</a></td>";
+        echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete Category')\" href='categories.php?delete={$admin_cat_id}'>Delete</a></td>";
+        // echo "<td><a href='categories.php?delete={$admin_cat_id}'>Delete</a></td>";
         echo "</tr>";
     }
 }
